pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peoplefinder-backend"
version = "1.0.0"
description = "FastAPI backend for PeopleFinder app"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "httpx==0.25.2",
    "aiohttp==3.9.1",
    "google-generativeai==0.3.0",
    "anthropic==0.7.0",
    "openai==1.3.9",
]
